Removed empty files+added CONSTANTS module
